LayoutTests/imported/w3c/web-platform-tests/credential-management/idlharness.https.window-expected.txt


PASS idl_test setup 
PASS idl_test validation 
PASS Partial interface Navigator: original interface defined 
PASS Partial interface Navigator: member names are unique 
PASS Partial dictionary CredentialRequestOptions: original dictionary defined 
PASS Partial dictionary CredentialRequestOptions: member names are unique 
PASS Partial dictionary CredentialCreationOptions: original dictionary defined 
PASS Partial dictionary CredentialCreationOptions: member names are unique 
PASS Partial dictionary CredentialRequestOptions[2]: original dictionary defined 
PASS Partial dictionary CredentialRequestOptions[2]: member names are unique 
PASS Partial dictionary CredentialCreationOptions[2]: original dictionary defined 
PASS Partial dictionary CredentialCreationOptions[2]: member names are unique 
PASS Partial interface mixin NavigatorID: member names are unique 
PASS PasswordCredential includes CredentialUserData: member names are unique 
PASS FederatedCredential includes CredentialUserData: member names are unique 
PASS HTMLElement includes GlobalEventHandlers: member names are unique 
PASS HTMLElement includes DocumentAndElementEventHandlers: member names are unique 
PASS HTMLElement includes ElementContentEditable: member names are unique 
PASS HTMLElement includes HTMLOrSVGElement: member names are unique 
PASS Navigator includes NavigatorID: member names are unique 
PASS Navigator includes NavigatorLanguage: member names are unique 
PASS Navigator includes NavigatorOnLine: member names are unique 
PASS Navigator includes NavigatorContentUtils: member names are unique 
PASS Navigator includes NavigatorCookies: member names are unique 
PASS Navigator includes NavigatorPlugins: member names are unique 
PASS Navigator includes NavigatorConcurrentHardware: member names are unique 
PASS Element includes ParentNode: member names are unique 
PASS Element includes NonDocumentTypeChildNode: member names are unique 
PASS Element includes ChildNode: member names are unique 
PASS Element includes Slotable: member names are unique 
PASS Credential interface: existence and properties of interface object 
PASS Credential interface object length 
PASS Credential interface object name 
PASS Credential interface: existence and properties of interface prototype object 
PASS Credential interface: existence and properties of interface prototype object's "constructor" property 
PASS Credential interface: existence and properties of interface prototype object's @@unscopables property 
PASS Credential interface: attribute id 
PASS Credential interface: attribute type 
PASS CredentialsContainer interface: existence and properties of interface object 
PASS CredentialsContainer interface object length 
PASS CredentialsContainer interface object name 
PASS CredentialsContainer interface: existence and properties of interface prototype object 
PASS CredentialsContainer interface: existence and properties of interface prototype object's "constructor" property 
PASS CredentialsContainer interface: existence and properties of interface prototype object's @@unscopables property 
PASS CredentialsContainer interface: operation get(optional CredentialRequestOptions) 
PASS CredentialsContainer interface: operation store(Credential) 
PASS CredentialsContainer interface: operation create(optional CredentialCreationOptions) 
PASS CredentialsContainer interface: operation preventSilentAccess() 
PASS CredentialsContainer must be primary interface of navigator.credentials 
PASS Stringification of navigator.credentials 
PASS CredentialsContainer interface: navigator.credentials must inherit property "get(optional CredentialRequestOptions)" with the proper type 
PASS CredentialsContainer interface: calling get(optional CredentialRequestOptions) on navigator.credentials with too few arguments must throw TypeError 
PASS CredentialsContainer interface: navigator.credentials must inherit property "store(Credential)" with the proper type 
PASS CredentialsContainer interface: calling store(Credential) on navigator.credentials with too few arguments must throw TypeError 
PASS CredentialsContainer interface: navigator.credentials must inherit property "create(optional CredentialCreationOptions)" with the proper type 
PASS CredentialsContainer interface: calling create(optional CredentialCreationOptions) on navigator.credentials with too few arguments must throw TypeError 
PASS CredentialsContainer interface: navigator.credentials must inherit property "preventSilentAccess()" with the proper type 
FAIL PasswordCredential interface: existence and properties of interface object assert_own_property: self does not have own property "PasswordCredential" expected property "PasswordCredential" missing
FAIL PasswordCredential interface object length assert_own_property: self does not have own property "PasswordCredential" expected property "PasswordCredential" missing
FAIL PasswordCredential interface object name assert_own_property: self does not have own property "PasswordCredential" expected property "PasswordCredential" missing
FAIL PasswordCredential interface: existence and properties of interface prototype object assert_own_property: self does not have own property "PasswordCredential" expected property "PasswordCredential" missing
FAIL PasswordCredential interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "PasswordCredential" expected property "PasswordCredential" missing
FAIL PasswordCredential interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "PasswordCredential" expected property "PasswordCredential" missing
FAIL PasswordCredential interface: attribute password assert_own_property: self does not have own property "PasswordCredential" expected property "PasswordCredential" missing
FAIL PasswordCredential interface: attribute name assert_own_property: self does not have own property "PasswordCredential" expected property "PasswordCredential" missing
FAIL PasswordCredential interface: attribute iconURL assert_own_property: self does not have own property "PasswordCredential" expected property "PasswordCredential" missing
FAIL PasswordCredential must be primary interface of passwordCredential assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: passwordCredential"
FAIL Stringification of passwordCredential assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: passwordCredential"
FAIL PasswordCredential interface: passwordCredential must inherit property "password" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: passwordCredential"
FAIL PasswordCredential interface: passwordCredential must inherit property "name" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: passwordCredential"
FAIL PasswordCredential interface: passwordCredential must inherit property "iconURL" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: passwordCredential"
FAIL Credential interface: passwordCredential must inherit property "id" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: passwordCredential"
FAIL Credential interface: passwordCredential must inherit property "type" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: passwordCredential"
FAIL FederatedCredential interface: existence and properties of interface object assert_own_property: self does not have own property "FederatedCredential" expected property "FederatedCredential" missing
FAIL FederatedCredential interface object length assert_own_property: self does not have own property "FederatedCredential" expected property "FederatedCredential" missing
FAIL FederatedCredential interface object name assert_own_property: self does not have own property "FederatedCredential" expected property "FederatedCredential" missing
FAIL FederatedCredential interface: existence and properties of interface prototype object assert_own_property: self does not have own property "FederatedCredential" expected property "FederatedCredential" missing
FAIL FederatedCredential interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "FederatedCredential" expected property "FederatedCredential" missing
FAIL FederatedCredential interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "FederatedCredential" expected property "FederatedCredential" missing
FAIL FederatedCredential interface: attribute provider assert_own_property: self does not have own property "FederatedCredential" expected property "FederatedCredential" missing
FAIL FederatedCredential interface: attribute protocol assert_own_property: self does not have own property "FederatedCredential" expected property "FederatedCredential" missing
FAIL FederatedCredential interface: attribute name assert_own_property: self does not have own property "FederatedCredential" expected property "FederatedCredential" missing
FAIL FederatedCredential interface: attribute iconURL assert_own_property: self does not have own property "FederatedCredential" expected property "FederatedCredential" missing
FAIL FederatedCredential must be primary interface of federatedCredential assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: federatedCredential"
FAIL Stringification of federatedCredential assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: federatedCredential"
FAIL FederatedCredential interface: federatedCredential must inherit property "provider" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: federatedCredential"
FAIL FederatedCredential interface: federatedCredential must inherit property "protocol" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: federatedCredential"
FAIL FederatedCredential interface: federatedCredential must inherit property "name" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: federatedCredential"
FAIL FederatedCredential interface: federatedCredential must inherit property "iconURL" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: federatedCredential"
FAIL Credential interface: federatedCredential must inherit property "id" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: federatedCredential"
FAIL Credential interface: federatedCredential must inherit property "type" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: federatedCredential"
PASS Navigator interface: attribute credentials 
